데코레이터 패턴 수정

+++ src/main/java/uknowlab/designpattern/decorator/beverage/Beverage.java
@@ -7,5 +7,5 @@ public abstract class Beverage {
         return description;
     }
 
-    public abstract double cost();
+    public abstract int cost();
 }
